Split verification logic to separate module
