+++ crates/sentinel/src/sentinel/mod.rs
@@ -1,7 +1,9 @@
 pub mod handler;
 pub mod listener;
 pub mod poller;
+pub mod verification;
 
 pub use handler::Sentinel;
 pub use listener::ReqListener;
 pub use poller::PollingSentinel;
+pub use verification::ValidatorVerifier;
